Own Project: Please Rubocop in Lanalytics

See merge request xikolo/lanalytics!314

+++ db/migrate/20160227170004_rebuild_cluster_groups.rb
@@ -1,8 +1,8 @@
+# frozen_string_literal: true
+
 class RebuildClusterGroups < ActiveRecord::Migration[4.2]
   def change
-    if ActiveRecord::Base.connection.table_exists? 'cluster_groups'
-      drop_table :cluster_groups
-    end
+    drop_table :cluster_groups if ActiveRecord::Base.connection.table_exists? 'cluster_groups' # rubocop:disable Rails/ReversibleMigration
 
     create_table :cluster_groups, id: :uuid do |t|
       t.string :name
